separate files for nickname functions
